+++ library/src/lapack/roclapack_sytrf_strided_batched.cpp
@@ -1,0 +1,124 @@
+/* ************************************************************************
+ * Copyright (c) 2019-2021 Advanced Micro Devices, Inc.
+ * ************************************************************************ */
+
+#include "roclapack_sytrf.hpp"
+
+template <typename T, typename U>
+rocblas_status rocsolver_sytrf_strided_batched_impl(rocblas_handle handle,
+                                                    const rocblas_fill uplo,
+                                                    const rocblas_int n,
+                                                    U A,
+                                                    const rocblas_int lda,
+                                                    const rocblas_stride strideA,
+                                                    rocblas_int* ipiv,
+                                                    const rocblas_stride strideP,
+                                                    rocblas_int* info,
+                                                    const rocblas_int batch_count)
+{
+    ROCSOLVER_ENTER_TOP("sytrf_strided_batched", "--uplo", uplo, "-n", n, "--lda", lda, "--strideA",
+                        strideA, "--strideP", strideP, "--batch_count", batch_count);
+
+    if(!handle)
+        return rocblas_status_invalid_handle;
+
+    // argument checking
+    rocblas_status st
+        = rocsolver_sytf2_sytrf_argCheck(handle, uplo, n, lda, A, ipiv, info, batch_count);
+    if(st != rocblas_status_continue)
+        return st;
+
+    // using unshifted arrays
+    rocblas_int shiftA = 0;
+
+    // memory workspace sizes:
+    // size of reusable workspace
+    size_t size_work;
+    rocsolver_sytrf_getMemorySize<T>(n, batch_count, &size_work);
+
+    if(rocblas_is_device_memory_size_query(handle))
+        return rocblas_set_optimal_device_memory_size(handle, size_work);
+
+    // memory workspace allocation
+    void* work;
+    rocblas_device_malloc mem(handle, size_work);
+
+    if(!mem)
+        return rocblas_status_memory_error;
+
+    work = mem[0];
+
+    // execution
+    return rocsolver_sytrf_template<T>(handle, uplo, n, A, shiftA, lda, strideA, ipiv, strideP,
+                                       info, batch_count, (T*)work);
+}
+
+/*
+ * ===========================================================================
+ *    C wrapper
+ * ===========================================================================
+ */
+
+extern "C" {
+
+rocblas_status rocsolver_ssytrf_strided_batched(rocblas_handle handle,
+                                                const rocblas_fill uplo,
+                                                const rocblas_int n,
+                                                float* A,
+                                                const rocblas_int lda,
+                                                const rocblas_stride strideA,
+                                                rocblas_int* ipiv,
+                                                const rocblas_stride strideP,
+                                                rocblas_int* info,
+                                                const rocblas_int batch_count)
+{
+    return rocsolver_sytrf_strided_batched_impl<float>(handle, uplo, n, A, lda, strideA, ipiv,
+                                                       strideP, info, batch_count);
+}
+
+rocblas_status rocsolver_dsytrf_strided_batched(rocblas_handle handle,
+                                                const rocblas_fill uplo,
+                                                const rocblas_int n,
+                                                double* A,
+                                                const rocblas_int lda,
+                                                const rocblas_stride strideA,
+                                                rocblas_int* ipiv,
+                                                const rocblas_stride strideP,
+                                                rocblas_int* info,
+                                                const rocblas_int batch_count)
+{
+    return rocsolver_sytrf_strided_batched_impl<double>(handle, uplo, n, A, lda, strideA, ipiv,
+                                                        strideP, info, batch_count);
+}
+
+rocblas_status rocsolver_csytrf_strided_batched(rocblas_handle handle,
+                                                const rocblas_fill uplo,
+                                                const rocblas_int n,
+                                                rocblas_float_complex* A,
+                                                const rocblas_int lda,
+                                                const rocblas_stride strideA,
+                                                rocblas_int* ipiv,
+                                                const rocblas_stride strideP,
+                                                rocblas_int* info,
+                                                const rocblas_int batch_count)
+{
+    return rocsolver_sytrf_strided_batched_impl<rocblas_float_complex>(
+        handle, uplo, n, A, lda, strideA, ipiv, strideP, info, batch_count);
+}
+
+rocblas_status rocsolver_zsytrf_strided_batched(rocblas_handle handle,
+                                                const rocblas_fill uplo,
+                                                const rocblas_int n,
+                                                rocblas_double_complex* A,
+                                                const rocblas_int lda,
+                                                const rocblas_stride strideA,
+                                                rocblas_int* ipiv,
+                                                const rocblas_stride strideP,
+                                                rocblas_int* info,
+                                                const rocblas_int batch_count)
+{
+    return rocsolver_sytrf_strided_batched_impl<rocblas_double_complex>(
+        handle, uplo, n, A, lda, strideA, ipiv, strideP, info, batch_count);
+}
+
+} // extern C