+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Full documentation for rocSOLVER is available at [rocsolver.readthedocs.io](http
 ### Added
 - Symmetric matrix factorizations:
     - LASYF
-    - SYTF2 (with batched and strided\_batched versions)
+    - SYTF2, SYTRF (with batched and strided\_batched versions)
 - Added `rocsolver_get_version_string_size` to help with version string queries
 
 ### Optimized